Add Swagger Correct invalid keycloak test json Move Bulk Update package and tidy
