+++ src/main/java/uk/gov/digital/ho/hocs/info/bulkupdate/BulkImportException.java
@@ -1,7 +1,11 @@
-package uk.gov.digital.ho.hocs.info.security;
+package uk.gov.digital.ho.hocs.info.bulkupdate;
 
 public class BulkImportException extends RuntimeException {
     public BulkImportException(String message, Exception cause) {
         super(message, cause);
+    }
+
+    public BulkImportException(String message) {
+        super(message);
     }
 }
