+++ src/main/resources/db/postgresql/V1_99__SEED_DATA.sql
@@ -84,7 +84,26 @@ INSERT INTO unit_case_type (unit_uuid, case_type)
 VALUES ('d9a93c21-a1a8-4a5d-aa7b-597bb95a782c', 'MIN');
 
 INSERT INTO permission (team_uuid, case_type, access_level)
-VALUES ('44444444-2222-2222-2222-222222222222', 'MIN', 'OWNER');
+VALUES ('44444444-2222-2222-2222-222222222222', 'MIN', 'OWNER'),
+       ('11111111-1111-1111-1111-111111111111', 'MIN', 'READ'),
+       ('33333333-3333-3333-3333-333333333333', 'MIN', 'WRITE'),
+       ('44444444-2222-2222-2222-222222222222', 'TRO', 'OWNER'),
+       ('11111111-1111-1111-1111-111111111111', 'TRO', 'READ'),
+       ('33333333-3333-3333-3333-333333333333', 'TRO', 'WRITE'),
+       ('44444444-2222-2222-2222-222222222222', 'DTEN', 'OWNER'),
+       ('11111111-1111-1111-1111-111111111111', 'DTEN', 'READ'),
+       ('33333333-3333-3333-3333-333333333333', 'DTEN', 'WRITE'),
+       ('44444444-2222-2222-2222-222222222222', 'IMCB', 'OWNER'),
+       ('11111111-1111-1111-1111-111111111111', 'IMCB', 'READ'),
+       ('33333333-3333-3333-3333-333333333333', 'IMCB', 'WRITE'),
+       ('44444444-2222-2222-2222-222222222222', 'IMCM', 'OWNER'),
+       ('11111111-1111-1111-1111-111111111111', 'IMCM', 'READ'),
+       ('33333333-3333-3333-3333-333333333333', 'IMCM', 'WRITE'),
+       ('44444444-2222-2222-2222-222222222222', 'UTEN', 'OWNER'),
+       ('11111111-1111-1111-1111-111111111111', 'UTEN', 'READ'),
+       ('33333333-3333-3333-3333-333333333333', 'UTEN', 'WRITE'),
+       ('44444444-2222-2222-2222-222222222222', 'UTEN', 'OWNER');
+
 
 INSERT INTO parent_topic (display_name, UUID)
 VALUES ('Parent topic 1', '11111111-1111-1111-1111-111111111121'),
@@ -218,7 +237,6 @@ VALUES ('3cb5d7ca-2198-4444-afd8-4a870f2d0b12',
 
 Insert INTO nominated_person (team_uuid, email_address)
 VALUES ('44444444-2222-2222-2222-222222222222', 'edward.liddiard@homeoffice.gsi.gov.uk'),
-       ('44444444-2222-2222-2222-222222222222', 'edward.liddiard@homeoffice.gsi.gov.uk'),
        ('11111111-1111-1111-1111-111111111111', 'edward.liddiard@homeoffice.gsi.gov.uk'),
        ('33333333-3333-3333-3333-333333333333', 'edward.liddiard@homeoffice.gsi.gov.uk');
